More seperation of interface and state
